[ssci] Added Shipped field to READMEs
This CL adds the Shipped field (and may update the
License File field) in Chromium READMEs. Changes were
automatically created, so if you disagree with any of
them (e.g. a package is used only for testing purposes
and is not shipped), comment the suggested change and
why.

See the LSC doc at go/lsc-chrome-metadata.

Bug: b:285450740
Change-Id: Id94e350df8f551d831b18f2270c0481210b79f91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/4663835
Reviewed-by: Frank Liberato <liberato@chromium.org>